pyproject.toml
[tool.poetry]
name = "assignment_db"
version = "0.1.0"
description = ""
authors = ["hajin <f-a-k@naver.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.1"
django-restframework = "^0.0.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
black = "^22.8.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
